pyproject.toml
[project]
name = "personal-agent-hub"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "feedparser==6.0.12",
    "httpx==0.28.1",
    "pydantic>=2.13.3",
    "pyyaml==6.0.3",
]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
    "ruff>=0.15.12",
]
